feat: move TaosStmt2BindTag implementation to bind.rs for better organization
